ㄹFIX: added test in FlowerShopService

+++ src/main/java/Unknown/PreScent/entity/SellerEntity.java
@@ -13,7 +13,7 @@ import javax.persistence.*;
 public class SellerEntity {
     @Id
     @Column(name = "seller_key")
-    private Long sellerKey;
+    private Integer sellerKey;
     @Column(name = "seller_name")
     private String sellerName;
     @Column(name = "seller_id")
